Fix OSSL_DECODER_CTX_set_selection not applying the selection value in FIPS.
Fixed the OSSL_DECODER_CTX_set_selection not applying the selection value
properly in the FIPS case.

Changed the selection of the struct decoder_pkey_data_st to the selection, that
is the pointer of the selection, because the selectionp can access the value
set by OSSL_DECODER_CTX_set_selection.